+++ R/style_tt.R
@@ -41,6 +41,7 @@
 #' + Can be combined such as: "lbt" to draw borders at the left, bottom, and top.
 #' @param line_color Color of the line. See the `color` argument for details.
 #' @param line_width Width of the line in em units (default: 0.1).
+#' @param finalize A function applied to the table object at the very end of table-building, for post-processing. For example, the function could use regular expressions to add LaTeX commands to the text version of the table hosted in `x@table_string`, or it could programmatically change the caption in `x@caption`.
 #' @param bootstrap_class String. A Bootstrap table class such as `"table"`, `"table table-dark"` or `"table table-dark table-hover"`. See the bootstrap documentation. 
 #' @param bootstrap_css A vector of CSS style declarations to be applied (ex: `"font-weight: bold"`). Each element corresponds to a cell defined by `i` and `j`.
 #' @param bootstrap_css_rule A string with complete CSS rules that apply to the table class specified using the `theme` argument of the `tt()` function.
@@ -50,22 +51,74 @@
 #' @return An object of class `tt` representing the table.
 #' @template latex_preamble
 #' @export
+#' @examplesIf knitr::is_html_output()
 #' @examples
+#' if (knitr::is_html_output()) options(tinytable_print_output = "html")
+#' 
 #' library(tinytable)
-#' x <- mtcars[1:5, 1:6]
-#' tab <- tt(x)
+#' 
+#' tt(mtcars[1:5, 1:6])
 #' 
 #' # Alignment
-#' style_tt(tab, j = 1:5, align = "lcccr")
-#' style_tt(tab, i = 2:3, background = "black", color = "orange", bold = TRUE)
-#' tab
+#' tt(mtcars[1:5, 1:6]) |> 
+#'   style_tt(j = 1:5, align = "lcccr")
+#' 
+#' # Colors and styles
+#' tt(mtcars[1:5, 1:6]) |> 
+#'   style_tt(i = 2:3, background = "black", color = "orange", bold = TRUE)
 #' 
 #' # column selection with `j``
-#' x <- mtcars[1:5, 1:6]
-#' tab <- tt(x)
-#' style_tt(tab, j = 5:6, background = "pink")
-#' style_tt(tab, j = "drat|wt", background = "pink")
-#' style_tt(tab, j = c("drat", "wt"), background = "pink")
+#' tt(mtcars[1:5, 1:6]) |> 
+#'   style_tt(j = 5:6, background = "pink")
+#' 
+#' tt(mtcars[1:5, 1:6]) |>
+#'   style_tt(j = "drat|wt", background = "pink")
+#' 
+#' tt(mtcars[1:5, 1:6]) |>
+#'   style_tt(j = c("drat", "wt"), background = "pink")
+#'
+#' tt(mtcars[1:5, 1:6], theme = "void") |>
+#'   style_tt(
+#'     i = 2, j = 2,
+#'     colspan = 3,
+#'     rowspan = 2,
+#'     align="c",
+#'     alignv = "m",
+#'     color = "white",
+#'     background = "black",
+#'     bold = TRUE)
+#'   
+#' tt(mtcars[1:5, 1:6], theme = "void") |>
+#'   style_tt(
+#'     i=0:3,
+#'     j=1:3,
+#'     line="tblr",
+#'     line_width=0.4,
+#'     line_color="teal")
+#'     
+#' tt(mtcars[1:5, 1:6], theme = "bootstrap") |>
+#'     style_tt(
+#'       i = c(2,5),
+#'       j = 3,
+#'       strikeout = TRUE,
+#'       fontsize = 0.7)
+#'       
+#' tt(mtcars[1:5, 1:6]) |>
+#'   style_tt(bootstrap_class = "table table-dark table-hover")
+#'
+#'
+#' inner <- "
+#' column{1-4}={halign=c},
+#' hlines = {fg=white},
+#' vlines = {fg=white},
+#' cell{1,6}{odd} = {bg=teal7},
+#' cell{1,6}{even} = {bg=green7},
+#' cell{2,4}{1,4} = {bg=red7},
+#' cell{3,5}{1,4} = {bg=purple7},
+#' cell{2}{2} = {r=4,c=2}{bg=azure7},
+#' "
+#' tt(mtcars[1:5, 1:4], theme = "void") |>
+#'   style_tt(tabularray_inner = inner)
 #'
 #' style_tt(tab,
 #'   i = 2, j = 2,
@@ -131,6 +184,7 @@ style_tt <- function (x,
                       line = NULL,
                       line_color = "black",
                       line_width = 0.1,
+                      finalize = NULL,
                       tabularray_inner = NULL,
                       tabularray_outer = NULL,
                       bootstrap_class = NULL,
@@ -171,6 +225,11 @@ style_tt <- function (x,
     out <- eval(cal)
   } else {
     out@lazy_style <- c(out@lazy_style, list(cal))
+  }
+
+  assert_function(finalize, null.ok = TRUE)
+  if (is.function(finalize)) {
+    out@lazy_finalize <- c(out@lazy_finalize, list(finalize))
   }
 
   return(out)
@@ -240,6 +299,11 @@ style_tt_lazy <- function (x,
       msg <- "`align` must be characters c, l, r, or d."
       stop(msg, call. = FALSE)
     }
+
+    if (any(align == "d")) {
+      tmp <- paste(sprintf("row{%s}={guard},", seq_len(x@nhead)), collapse = "\n")
+      tabularray_inner <- paste(tabularray_inner, tmp)
+    }
   }
 
   assert_style_tt(
@@ -253,6 +317,7 @@ style_tt_lazy <- function (x,
   if (!is.null(width)) {
     width <- paste0(width, "em")
   }
+
 
   out <- style_eval(x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout, color = color, background = background, fontsize = fontsize, width = width, align = align, alignv = alignv, colspan = colspan, rowspan = rowspan, indent = indent, tabularray_inner = tabularray_inner, tabularray_outer = tabularray_outer, bootstrap_css = bootstrap_css, bootstrap_css_rule = bootstrap_css_rule, bootstrap_class = bootstrap_class, line = line, line_color = line_color, line_width = line_width)