+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## Development
 
+
+New function `theme_tt()`:
+
 `format_tt()` improvements:
 
 * New `i` argument to format subsets of rows.
@@ -19,6 +22,7 @@ Typst format:
 
 Misc:
 
+* `style_tt()` gains a `finalize` argument. This accepts functions to be applied to the table object at the very end of the building process, to programmatically change its content. For example, this can be used with regular expressions to modify the text version of the table hosted in `tab@table_string`, or the function could programmatically modify the caption in `tab@caption`.
 * `style_tt()`: LaTeX format supports decimal alignement with `align="d"`. The width of columns is determined by the maximum number of digits to the left and to the right in all cells specified by `i`, `j`.
 * Support RevealJS slides in Quarto documents.
 * Improved support for `tibble`. ANSI characters (ex: fancy `pillar` formatting) are stripped automatically or converted to HTML when the `fansi` package is installed. `fansi` is a dependency of `tibble`, so it should often be installed.
@@ -29,9 +33,9 @@ Misc:
 
 Breaking changes:
 
-* In some cases, `format_tt()` could previously be applied sequentially to apply two formats to the same cell. Now, multiple calls to `format_tt()` can still be make chained with pipes, but they must apply to different cells with `i`, `j`, otherwise only the last change is respected. One exception is the `escape` argument which can be applied to pre-formatted cells.
-* The "tinytable_tabularray_placement" global option is renamed to "tinytable_tt_placement".
+* In some cases, `format_tt()` could be use sequentially to apply two formats to the same cell. Now, multiple calls to `format_tt()` can still be make chained with pipes, but they must apply to different cells with `i`, `j`, otherwise only the last change is respected. One exception is the `escape` argument which can be applied to pre-formatted cells.
 * `tinytable` objects no longer have a `meta_tinytable` attribute. Use S4 slots instead.
+* `placement` argument in `tt()` is removed in favor of `theme_tt("placement")`.
 
 Bugs: